requirements.txt
# Core dependencies
pillow>=9.0.0
python-magic>=0.4.24

# Web framework
flask>=2.3.0
werkzeug>=2.3.0

# Image processing and analysis
numpy>=1.20.0
opencv-python>=4.5.0
rawpy>=0.17.0

# Machine learning
scikit-learn>=1.0.0

# Command-line interface
click>=8.0.0
rich>=10.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
mypy>=0.900
black>=22.0.0